+++ compiler/src/main/java/info/teksol/mindcode/compiler/functions/UbindFunctionHandler.java
@@ -28,7 +28,7 @@ class UbindFunctionHandler extends AbstractFunctionHandler {
     }
 
     @Override
-    public String generateCall(List<NamedParameter> arguments) {
+    protected String generateCall(List<NamedParameter> arguments, boolean markOptional) {
         return getName() + "(" + BaseFunctionMapper.joinNamedArguments(arguments) + ")";
     }
 }